### 2025-06-18 #8424 @editorialbot
[PRE REVIEW]: Copernicus Seasonal Forecast Tools Package: Bridging Seasonal Climate Predictions and Impact Models for Operational Risk Assessment
**Submitting author:** @DahyannAraya (<a href="http://orcid.org/0000-0002-2439-6308">Dahyann Araya</a>)
**Repository:** https://github.com/DahyannAraya/copernicus-seasonal-forecast-tools
**Branch with paper.md** (empty if default branch):
**Version:** 0.1.1
