+++ paper.bib
@@ -1,181 +1,159 @@
 @article{AznarSiguan:2019,
+  url = {https://gmd.copernicus.org/articles/12/3085/2019/},
   author = {Aznar-Siguan, Gabriela and Bresch, David N.},
-  title = {{CLIMADA} v1: A global weather and climate risk assessment platform},
   journal = {Geoscientific Model Development},
+  keywords = {climate risk, hazard, vulnerability, adaptation},
+  month = jul,
+  title = {{CLIMADA} v1: A global weather and climate risk assessment platform},
+  year = {2019},
   volume = {12},
   number = {7},
   pages = {3085--3097},
-  year = {2019},
-  month = jul,
-  doi = {10.5194/gmd-12-3085-2019},
-  url = {https://gmd.copernicus.org/articles/12/3085/2019/},
-  abstract = {The open-source CLIMADA software integrates hazard, exposure, and vulnerability to assess global risk and compute socio-economic impacts. Demonstrated on Hurricane Irma 2017.},
-  language = {en},
-  keywords = {climate risk, hazard, vulnerability, adaptation}
+  doi = {10.5194/gmd-12-3085-2019}
 }
 
 @article{Geiger:2024,
+  url = {https://www.frontiersin.org/articles/10.3389/fclim.2024.1343993/full},
   author = {Geiger, Tobias and Röösli, Thomas and Bresch, David N. and Erhardt, Bodo and Fischer, Andreas M. and Imgrüth, Dominik and Kienberger, Stefan and Mainetti, Laura and Mühlbacher, Gudrun and Spiekermann, Raphael},
-  title = {How to provide actionable information on weather and climate impacts?–A summary of strategic, methodological, and technical perspectives},
   journal = {Frontiers in Climate},
+  keywords = {impact forecasting, NMHS, weather services, risk modeling},
+  month = mar,
+  title = {How to provide actionable information on weather and climate impacts?–A summary of strategic, methodological, and technical perspectives},
+  year = {2024},
   volume = {6},
   pages = {1343993},
-  year = {2024},
-  month = mar,
-  doi = {10.3389/fclim.2024.1343993},
-  url = {https://www.frontiersin.org/articles/10.3389/fclim.2024.1343993/full},
-  abstract = {This paper proposes a harmonized approach for National Meteorological and Hydrological Services (NMHSs) to provide actionable impact-related climate and weather services, using a risk framework that integrates hazard, exposure, and vulnerability.},
-  keywords = {impact forecasting, NMHS, weather services, risk modeling}
+  doi = {10.3389/fclim.2024.1343993}
 }
 
 @article{Osman:2023,
+  url = {https://doi.org/10.1016/j.cliser.2023.100397},
   author = {Osman, Marisol and Domeisen, Daniela I.V. and Robertson, Andrew W. and Weisheimer, Antje},
-  title = {Sub-seasonal to decadal predictions in support of climate services},
   journal = {Climate Services},
+  keywords = {sub-seasonal predictions, decadal predictions, climate services, forecast skill, climate variability, decision-making},
+  month = jun,
+  title = {Sub-seasonal to decadal predictions in support of climate services},
+  year = {2023},
   volume = {30},
   pages = {100397},
-  year = {2023},
-  month = jun,
-  doi = {10.1016/j.cliser.2023.100397},
-  url = {https://doi.org/10.1016/j.cliser.2023.100397},
[…]
