### 2025-06-21 4545c95
Update date format paper.md
