### 2025-06-22 #8424 @DahyannAraya
Thank you for your time and message, @kthyng.
This is my first time using this format; it seems to work on my end via GitHub Actions, but not on your side.
I’ve reviewed the error and updated the paper.md file.
I then triggered the PDF rebuild using @editorialbot generate pdf, and it appears to have worked :)

Please let me know if any further changes are needed. Have a nice week. Dahyann
