### 2025-07-07 1493503
generalize index calculation to flexible time window
