### 2025-07-07 82a1822
adapt tests
