### 2025-07-12 #8424 @elbeejay
@znicholls would you consider reviewing this submission titled: _Copernicus Seasonal Forecast Tools Package: Bridging Seasonal Climate Predictions and Impact Models for Operational Risk Assessment_, for JOSS? As someone with previous experience both publishing in JOSS and reviewing for JOSS, as well as expertise in both climate science and programming, you seem like an ideal reviewer for this submission.

Please let us know if you are able and interested in conducting a review. We are asking that reviewers complete their reviews within 6 weeks of us formally beginning the review issue. We can be flexible on that timeline as needed.

Thanks!
Jay
