### 2025-07-16 #8424 @elbeejay
@vergauwenthomas and @bradyrx, would either of you be willing to review the submission: _Copernicus Seasonal Forecast Tools Package: Bridging Seasonal Climate Predictions and Impact Models for Operational Risk Assessment_, for JOSS? I know you've both published Python climate-related packages in JOSS before, so in addition to being experts in the field, you are familar with our open review process.

Please let us know if you are able and interested in conducting a review. We are asking that reviewers complete their reviews within 6 weeks of us formally beginning the review issue. We can be flexible on that timeline as needed.

Thanks!
Jay
