### 2025-07-22 #8424 @elbeejay
@znicholls are you willing and able to review this submission: _"Copernicus Seasonal Forecast Tools Package: Bridging Seasonal Climate Predictions and Impact Models for Operational Risk Assessment"_, for JOSS? You might be a good fit given your background in climate modeling and Python. I know you've reviewed for JOSS in the past, so you are likely familiar with our process. At this time we're asking our reviewers to complete their initial reviews within 6 weeks of the official review thread beginning. If the material seems like a reasonable fit for you and your schedule permits, we would appreciate your involvement.

Thanks!
Jay
