### 2025-07-24 b3df4c7
Update README, Copernicus forecast notebook, and modules.rst
