+++ README.md
@@ -110,7 +110,7 @@ You can find further material in [<img src="https://colab.research.google.com/as
 
 ## **Resources**
 - [U-CLIMADAPT Project](https://www.copernicus-user-uptake.eu/user-uptake/details/responding-to-the-impact-of-climate-change-u-climadapt-488)
-- [Copernicus Seasonal Forecast on CLIMADA](to do)
+- [Copernicus Seasonal Forecas Tools package on CLIMADA](https://climada-python.readthedocs.io/en/latest/user-guide/climada_hazard_copernicus_seasonal_forecast.html)
 - [Copernicus Seasonal Forecast Tools package extended demostration](https://colab.research.google.com/github/DahyannAraya/climada_copernicus_seasonal_forecast_workshop/blob/main/DEMO_Modul_climada_copernicus_seasonal_forecast_workshop.ipynb)
 - [Seasonal forecast daily and subdaily data on single levels](https://cds.climate.copernicus.eu/datasets/seasonal-original-single-levels?tab=overview)
 - [Copernicus Climate Data Store](https://cds.climate.copernicus.eu)
