### 2025-07-25 b78ea28
Docs: fix RST issues and rebuild Sphinx sources
