+++ docs/source/autoapi/seasonal_forecast_tools/core/heat_index/index.rst
@@ -167,7 +167,7 @@ Module Contents
              * *Adjustments* --
 
                - If RH ≤ 13% and 80°F < T < 112°F:
-                   Adjustment = (13 - RH) / 4 * sqrt((17 - |T minus 95|) / 17)
+                   Adjustment = (13 - RH) / 4 * sqrt((17 - |T|) / 17)
 
                - If RH > 85% and T < 87°F:
                    Adjustment = (RH - 85) / 10 * ((87 - T) / 5)
