### 2025-07-30 #8424 @elbeejay
@abigailsnyder - hello! Would you be able to review this submission, "Copernicus Seasonal Forecast Tools Package: Bridging Seasonal Climate Predictions and Impact Models for Operational Risk Assessment", for JOSS? I know you've published in JOSS before so you are likely familiar with the process, please take a look at the submission materials above and let me know if you'll be able to help out on this one.

Best,
Jay
