### 2025-08-03 8dccd56
Update README and modules.rst
