### 2025-08-03 4dd0322
Update link in citing.rst
