+++ docs/source/citing.rst
@@ -28,4 +28,4 @@ If you use the **Copernicus Seasonal Forecast Tools** in your work, please cite 
 
 You may also want to cite the related project:
 
-* U-CLIMADAPT Project (2023-2025). Responding to the impact of climate change. Copernicus User Uptake Framework Partnership Agreement. https://www.copernicus-user-uptake.eu/user-uptake/details/responding-to-the-impact-of-climate-change-u-climadapt-488
+* U-CLIMADAPT Project (2023-2025). Responding to the impact of climate change. Copernicus User Uptake Framework Partnership Agreement. https://www.dwd.de/EN/research/projects/socioeconomics/fpcup_u_climadapt/fpcup_u_climadapt.html
