### 2025-08-09 #8424 @falconstryker
@elbeejay apologies for my late reply, I just returned from traveling abroad. I'm happy to review this software for JOSS. This will be my first JOSS review so I appreciate any tips you have. I do know where to find the reviewer instructions online.
