### 2025-08-15 #8424 @elbeejay
@DamienIrving - wondering if you'll consider reviewing _"Copernicus Seasonal Forecast Tools Package: Bridging Seasonal Climate Predictions and Impact Models for Operational Risk Assessment"_, for JOSS? I understand you are a climate scientist with expertise in Python, and have reviewed for JOSS in years past. Please let us know if you, or a colleague, might be able to help review this submission.

Thanks!
Jay
