### 2025-08-16 #8827 @elbeejay
@falconstryker - I know this is your first review for JOSS. The first few posts provide some additional information about the process as well as links to JOSS review documentation. Additional tips and suggestions I'd offer are the following:

- Unlike a traditional review process, JOSS is far more conversational; we do not have explicity review "rounds" and you can have more of a dialog and back-forth with the authors. Consequently, it can be good to start your review early, so you can flag any issues you run into immediately; I don't expect it to be an issue with this submission, but we've had reviewers run into errors when installing packages in the past and that obviously prevents them from conducting the rest of their review
- In that same spirit, you are able to comment on or otherwise reference issues or comments made by another reviewer; this can help emphasize comments and stress important points you see being made
- In the spirit of open communication, we try to be highly flexible with timelines; if you are going to be busy with other work or need an extension to complete your review that is no problem, I just ask that you let us know here so the submitting authors are not left wondering what is happening

Feel free to reach out should you have any further questions, otherwise the command for you to generate a reviewer checklist should be in the first comment in this issue.

--

@falconstryker and @DamienIrving many thanks again for agreeing to review for JOSS. We're trying to operate on a 6 week review timeline at the moment, so if it is possible for you to get through your initial reviews by September 27, that would be fantastic. Again, just let us know here if you need an extension, that is no problem.

Best,
Jay
