### 2025-08-21 #8827 @nishadhka
COI disclose, an abstract in EGU with one of the authors of the paper https://meetingorganizer.copernicus.org/EGU25/EGU25-15886.html
