### 2025-08-26 DahyannAraya/copernicus-seasonal-forecast-tools#4 @DamienIrving
Community guidelines
I'm reviewing the JOSS submission for this package (see https://github.com/openjournals/joss-reviews/issues/8827) and overall you've done a terrific job with this package.

The only thing that is possibly missing from the review checklist is community guidelines:
> Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support

At the documentation page I could find a short note on [installation instructions for developers/contributors](https://copernicus-seasonal-forecast-tools.readthedocs.io/en/latest/index.html#developer-setup), but nothing else with directions on how to contribute/modify (I can imagine some people might want to add their own indices), report issues or seek support.
