### 2025-08-26 #8827 @DamienIrving
I've read the manuscript and was able to successfully run the examples in the documentation - it's a great package.

The only issue I've identified is a lack of community guidelines, so I've created an issue to flag that: https://github.com/DahyannAraya/copernicus-seasonal-forecast-tools/issues/4

Well done to the authors!
