### 2025-08-27 DahyannAraya/copernicus-seasonal-forecast-tools#5 @falconstryker
Missing DOI
I'm reviewing the JOSS submission for this package (see [here](https://github.com/openjournals/joss-reviews/issues/8827)) and noticed one of your citations in the paper has a DOI that isn't listed:

For Murphy et al. (2000), Seasonal forecasting for climate hazards: Prospects and responses, the DOI is: 10.1023/a:1011160904414

Please substitute the URL in references for the DOI.
