### 2025-08-27 #8827 @falconstryker
I have started my review. So far, I have flagged just [one issue](https://github.com/DahyannAraya/copernicus-seasonal-forecast-tools/issues/5#issue-3361089059) with a citation and a missing definition in the software paper. I plan on starting the software functionality piece tomorrow.
