+++ paper.bib
@@ -32,7 +32,7 @@
   keywords = {sub-seasonal predictions, decadal predictions, climate services, forecast skill, climate variability, decision-making},
   month = jun,
   title = {Sub-seasonal to decadal predictions in support of climate services},
-  year = {2023},
+  year = {2023},s
   volume = {30},
   pages = {100397},
   doi = {10.1016/j.cliser.2023.100397}
@@ -119,7 +119,8 @@
   journal = {Climatic Change},
   keywords = {seasonal forecasting, climate hazards, risk management, policy response, institutional capacity},
   title = {Seasonal Forecasting for Climate Hazards: Prospects and Responses},
-  year = {2000}
+  year = {2000},
+  doi = {10.1023/a:1011160904414}
 }
 
 @article{Buontempo:2022,
