+++ paper.bib
@@ -32,7 +32,7 @@
   keywords = {sub-seasonal predictions, decadal predictions, climate services, forecast skill, climate variability, decision-making},
   month = jun,
   title = {Sub-seasonal to decadal predictions in support of climate services},
-  year = {2023},s
+  year = {2023},
   volume = {30},
   pages = {100397},
   doi = {10.1016/j.cliser.2023.100397}
