### 2025-09-22 DahyannAraya/copernicus-seasonal-forecast-tools#6 @nishadhka
Avoid sample inputs in deep nested folders
This nested test file organization can be avoided.

https://github.com/DahyannAraya/copernicus-seasonal-forecast-tools/tree/main/test/climada_data_test/copernicus_data/seasonal_forecasts/dwd/sys21/2022/init11/valid12_02/downloaded_data/grib/*.grib
