### 2025-09-22 DahyannAraya/copernicus-seasonal-forecast-tools#7 @nishadhka
Using well known method in utils
Instead of custom rewriting the routines, getting the bounding box from shapely or other well known libraries can be tried.  https://github.com/DahyannAraya/copernicus-seasonal-forecast-tools/blob/54c2b24101b93d2cbba5a861b6eb6891c1e70abd/seasonal_forecast_tools/utils/coordinates_utils.py#L58

Such as https://shapely.readthedocs.io/en/2.1.1/reference/shapely.box.html
