### 2025-09-24 #8827 @falconstryker
@elbeejay I have a question, but I need to preface it first: the review is for the copernicus-seasonal-forecast-tools repo in DahyannAraya's GitHub, but I see from this project's history that this is a forked repo from the **Weather and Climate Risks Research group at ETH Zurich, who's GitHub is CLIMADA-project (https://github.com/CLIMADA-project/copernicus-seasonal-forecast-tools)**.

My question is: Do the authors want the official review to be on DahyannAraya's GitHub or CLIMADA-project?

I ask because I have my own project queued for JOSS review and the author's situation here might be similar to mine: we want our official release under the NASA GitHub to be reviewed, not my personal forked repo (which I just use to test new features and then open PR's to the official upstream repo). I want to flag this to make sure the authors get the review they're looking for, especially if they're hoping for a citable DOI for their official repo.
