### 2025-09-24 #8827 @falconstryker
Raising another question @elbeejay - David Bresch and Tobias Geiger are listed as authors in the JOSS paper, but I don't see their contributions to the code in the commit history. It is unclear how they've contributed to authorship - can you advise on next steps?
