### 2025-09-24 DahyannAraya/copernicus-seasonal-forecast-tools#4 @falconstryker
As a reviewer for JOSS (https://github.com/openjournals/joss-reviews/issues/8827), I'm commenting to suggest closing this thread if the issue has been addressed.
