### 2025-09-24 DahyannAraya/copernicus-seasonal-forecast-tools#8 @falconstryker
Two JOSS paper items
I'm working on your review for JOSS (https://github.com/openjournals/joss-reviews/issues/8827#issuecomment-3229945648) and have two paper items of note:

1. State of the field is incompletely addressed. The paper notes that the software addresses a gap in the data analysis process, but does not compare to other commonly used packages.
2. The Copernicus Climate Data Store (CDS) isn't cited in the paper (line 47).

Hopefully these are simple to address!
