### 2025-09-25 DahyannAraya/copernicus-seasonal-forecast-tools#9 @falconstryker
readthedocs organization and clarity
### Cleaning up readthedocs

Overall, the readthedocs site needs work.

The installation instructions need to be separate from the homepage and clearly associated with the Copernicus API page, not the table of contents or community guidelines. I recommend that the homepage contain a description of the software, community guidelines, table of contents, license, and obvious link to a separate page for installation.

The Copernicus API page should be nested under the installation page or included on the same page.

The API Reference page is confusing. I know from having used Sphinx before that its difficult to customize the readthedocs setup, but it is necessary in order for readthedocs to be a useful guide. I recommend renaming "API reference" to "detailed documentation" and then organizing the sublabels that appear under the page. Also clean up some content – for example, the [seasonal_forecast_tools.data.downloader](https://copernicus-seasonal-forecast-tools.readthedocs.io/en/latest/autoapi/seasonal_forecast_tools/data/downloader/index.html#module-seasonal_forecast_tools.data.downloader) page contains install instructions that don't belong there.

There should be just one page for the examples/walkthrough for the software. The sidebar currently shows 6 pages for this step, when really the 5 pages below "Copernicus Seasonal Forecast Tools Package" are headers on that page:

> Home
> CDS API
> API Reference
> Copernicus Seasonal Forecast Tools Package
> Set up parameters   _-> a header on the Copernicus Seasonal... page_
> Download and Process Data   _-> a header on the Copernicus Seasonal... page_
> Calculate Climate Indices   _-> a header on the Copernicus Seasonal... page_
> Calculate a Hazard Object   _-> a header on the Copernicus Seasonal... page_
> Example for reading and plotting hazard   _-> a header on the Copernicus Seasonal... page_
> How to Cite
> Resources

Combining my recommendations so far, the sidebar should look something more like this:
> Home
> Installation
> Detailed Documentation
> Copernicus Seasonal Forecast Tools Package Walkthrough
> How to Cite
> Resources

### Installation Clarifications
Be sure to specify that users can either do "pip install ." or "pip install -r docs/requirements.txt" but not both in the instructions (https://copernicus-seasonal-forecast-tools.readthedocs.io/en/latest/index.html):

> Install from GitHub:
> ```
> git clone https://github.com/DahyannAraya/copernicus-seasonal-forecast-tools.git
> cd copernicus-seasonal-forecast-tools
> pip install .
> pip install -r docs/requirements.txt  # optional
> ```

Also, note that there is a developer installation option earlier in the text. Right now it reads as if there is a PyPi, Conda/Mamba, and GitHub install only. The developer option should appear above with those three options. Also, please separate out instructions for Windows vs. Mac vs. Linux, e.g., the section quoted below is confusing for beginners:

> Developer Setup
> To contribute or run the documentation and tests:
>
> ```
> git clone https://github.com/DahyannAraya/copernicus-seasonal-forecast-tools.git
> cd copernicus-seasonal-forecast-tools
> python3.10 -m venv .venv_forecast  # Windows: use py -3.10
> source .venv_forecast/bin/activate
> pip install -e .
> pip install -r docs/requirements.txt
> ```
>
> Note: On Windows, use `.venv_forecast\Scripts\activate` instead of `source`.

### Walkthrough

On the Copernicus Seasonal Forecast Tools page, which appears to be a walkthrough/example page (https://copernicus-seasonal-forecast-tools.readthedocs.io/en/latest/climada_hazard_copernicus_forecast.html), there is a section for prerequisites:

> Prerequisites:
> 1. CDS account and API key: Register at https://cds.climate.copernicus.eu
> 2. CDS API client installation: pip install cdsapi
> 3. CDS API configuration: Create a .cdsapirc file in your home directory with your API key and URL. For instructions, visit: https://cds.climate.copernicus.eu/how-to-api#install-the-cds-api-client
> 4. Dataset Terms and Conditions: After selecting the dataset to download, make sure to accept the terms and conditions on the corresponding dataset webpage in the CDS portal before running this notebook. Here, https://cds.climate.copernicus.eu/datasets/seasonal-original-single-levels?tab=download.

that would be enhanced by including links to relevant parts of the readthedocs site.

### Main Problem:
The [Copernicus Seasonal Forecast Tools](https://copernicus-seasonal-forecast-tools.readthedocs.io/en/latest/climada_hazard_copernicus_forecast.html) page is hard to read and follow. It reads more like the detailed documentation that the autodoc tries to supply, but it has a bit of code mixed in as if its an example use page. It would be a great example use page if it were re-written to explain a whole example start-to-finish. I have played around with the code in python and I can see that the code provided here does showcase the utility of the software, but the text doesn't support that. **Please write a clear, comprehensive example/walkthrough for the user, and move the descriptions of the functions to the detailed documentation elsewhere.**
