### 2025-09-25 DahyannAraya/copernicus-seasonal-forecast-tools#10 @falconstryker
Broken links in readthedocs
There are several broken links in the readthedocs.

### On the page for setting up the Copernicus API (https://copernicus-seasonal-forecast-tools.readthedocs.io/en/latest/cds_api.html):

The line under step 3 that says:

> After registration, go to your CDS account page: https://cds.climate.copernicus.eu/api-how-to

 the link is broken and leads to a "page not found" error.

Same for the link under step 4 that says:

> Navigate to the dataset you want, for example: https://cds.climate.copernicus.eu/datasets/seasonal-original-single-levels?tab=download

### On the homepage (https://copernicus-seasonal-forecast-tools.readthedocs.io/en/latest/index.html):

The part of the sentence that says:

> part of the [U-CLIMADAPT](https://www.copernicus-user-uptake.eu/user-uptake/details/responding-to-the-impact-of-climate-change-u-climadapt-488) project

the U-CLIMADAPT link is broken

### On the Copernicus Seasonal Forecast Tools page (https://copernicus-seasonal-forecast-tools.readthedocs.io/en/latest/climada_hazard_copernicus_forecast.html)

There are 3 URLs that are not links:

> 3. CDS API configuration: Create a .cdsapirc file in your home directory with your API key and URL. For instructions, visit: https://cds.climate.copernicus.eu/how-to-api#install-the-cds-api-client
> 4. Dataset Terms and Conditions: After selecting the dataset to download, make sure to accept the terms and conditions on the corresponding dataset webpage in the CDS portal before running this notebook. Here, https://cds.climate.copernicus.eu/datasets/seasonal-original-single-levels?tab=download.
> ...
You can view the terms and conditions at https://cds.climate.copernicus.eu/datasets/seasonal-original-single-levels?tab=download.
