### 2025-09-26 3a8fd56
Add CONTRIBUTE.md, update README and html
