+++ CONTRIBUTE.md
@@ -1,0 +1,27 @@
+## **Community guidelines**
+
+This section summarizes how to contribute and where to find details. We adopt the CLIMADA contribution workflow and conventions.
+
+### 1) Contribute 
+We follow the CLIMADA workflow (branching, tests, reviews). Please read the CLIMADA guide:
+https://climada-python.readthedocs.io/en/latest/development/index.html  
+Typical steps: fork → branch → tests → pull request.
+For extended resorces for CLIMADA Development visit https://climada-python.readthedocs.io/en/latest/development/Guide_CLIMADA_Development.html
+
+### 2) Report issues or problems
+**[Open an issue](https://github.com/DahyannAraya/copernicus-seasonal-forecast-tools/issues/new/choose)**. Please include: a description, minimal example, expected vs. actual behavior, version info (package, Python, OS), and any logs/traceback.
+
+### 3) Seek support
+**[Talk to the CLIMADA group](https://climada-python.readthedocs.io/en/latest/development/Guide_CLIMADA_Development.html#talk-to-the-group)**
+- Before starting to code, coordinate with one of the repo admins.
+- Use this chance to align on the big picture with the group: intersections with other projects, potential conflicts, changes to CLIMADA core, and additional dependencies.
+- Also talk with members of the core development team (see the GitHub wiki).
+- Bring your idea to a developers meeting to get advice, find collaborators, and identify reviewers.
+- Keep communicating as your plans evolve.
+
+### 4) How to Stay Informed or Participate?
+
+- [GitHub repository](https://github.com/CLIMADA-project/climada_python)
+- [Developer and user mailing list](https://sympa.ethz.ch/sympa/info/climada-users)
+- Monthly developers' and users' meetings (please register for the mailing list to receive invitations).
+- [Contribution guide](https://climada-python.readthedocs.io/en/latest/development/index.html)
