### 2025-09-26 7034337
Update paper.md and bib
