+++ paper.md
@@ -46,11 +46,13 @@ The landscape of disaster risk management is undergoing a significant transforma
 
 A three-step conceptual framework illustrates the progression from prediction to actionable advice: 1) seasonal climate prediction, 2) impact modeling, and 3) actionable impact-based forecasting. While significant advances have been made in the first two steps separately, a critical gap remains in operational systems that systematically connect seasonal forecasts with impact models to enable the third step. To address this challenge, we introduce the Copernicus Seasonal Forecast  Package, a tool that integrates seasonal climate predictions into the CLIMADA risk modeling pipeline. This work helps bridge the strategic gap between traditional seasonal forecasts and impact-oriented services, contributing to the development of operational, reproducible workflows that connect probabilistic climate forecasts to warning systems and decision-making processes.
 
-The package automates the entire data flow—from data retrieval via the Copernicus Climate Data Store (CDS), through preprocessing and heat  index calculation, to the generation of CLIMADA-compatible hazard objects. This integration enables researchers and practitioners to transform seasonal climate forecasts into concrete impact assessments, supporting proactive risk management strategies months ahead of potential events.
+The package automates the entire data flow—from data retrieval via the Copernicus Climate Data Store [CDS], through preprocessing and heat  index calculation, to the generation of CLIMADA-compatible hazard objects. This integration enables researchers and practitioners to transform seasonal climate forecasts into concrete impact assessments, supporting proactive risk management strategies months ahead of potential events.
 
 # Statement of need
 
 Despite the growing interest in both seasonal prediction [@Osman:2023; @NgoungueLangue:2025; @Murphy:2000] and impact modelling — aiming to shift from describing what the hazard will look like to how the hazard will impact [@Merz:2020; @Shyrokaya:2024], and moving from weather-centric forecasts to impact-oriented services — there is still a lack of operational and reproducible tools to link probabilistic seasonal climate predictions with quantitative impact assessment frameworks [@Geiger:2024]. Current workflows are typically fragmented, requiring manual intervention across multiple stages including data access, preprocessing, and hazard transformation. These inefficiencies can delay critical risk assessments and compromise reproducibility [@Potter:2025; @Wyatt:2023].
+
+While existing libraries such as xclim [Bourgault:2023], icclim [Page:2022], and Thermofeel [Brimicombe:2022] provide robust implementations of climate indices, they do not offer an operational pathway from seasonal forecasts to impact modeling frameworks. Our work complements these established tools by focusing specifically on the operational gap between seasonal climate predictions and risk assessment. We utilize Thermofeel as a reference for specific index calculations, ensuring our formulas align with published definitions to maintain comparability and facilitate reuse across the scientific community.
 
 The Copernicus Seasonal Forecast Tools Package addresses this gap by providing an automated, flexible, and modular solution. It connects seasonal forecast data from the Copernicus Climate Data Store (CDS) with the hazard and impact modeling capabilities of the CLIMADA platform [@AznarSiguan:2019], an open-source platform for climate risk analysis and the evaluation of adaptation benefits [@bresch_climada_2021]. This integration allows users to efficiently transform raw seasonal predictions into actionable climate risk insights. The package supports the definition of custom climate indices, the integration of data from multiple providers, and the adaptation to emerging hazard types and regional contexts. Thereby, it strengthens the operational implementation of impact forecasting and facilitates downstream components of the warning value chain or operational pipeline [@Golding:2019], bridging the gap between probabilistic seasonal forecasts and concrete impact assessments in an actionable timescale.
 
