### 2025-09-26 676e7ce
Update paper.bib
