+++ paper.bib
@@ -150,6 +150,15 @@
   doi = {10.1029/2020RG000704}
 }
 
+@misc{CDS,
+  url = {https://cds.climate.copernicus.eu/},
+  author = {{Copernicus Climate Change Service}},
+  title = {Description of the {C3S} seasonal multi-system},
+  year = {2024},
+  howpublished = {\url{https://cds.climate.copernicus.eu/}},
+  note = {Accessed: 2024-12-01}
+}
+
 @misc{cds_documentation,
   url = {https://confluence.ecmwf.int/display/CKB/Description+of+the+C3S+seasonal+multi-system},
   author = {{Copernicus Climate Change Service}},
@@ -157,4 +166,36 @@
   year = {2023},
   howpublished = {\url{https://confluence.ecmwf.int/display/CKB/Description+of+the+C3S+seasonal+multi-system}},
   note = {Accessed: 2023-05-03}
+}
+
+@article{Bourgault:2023,
+  url = {https://joss.theoj.org/papers/10.21105/joss.05415},
+  title = "xclim: xarray-based climate data analytics",
+  author = "Bourgault, Pascal and Huard, David and Smith, Trevor James and Logan, Travis and others",
+  journal = "Journal of Open Source Software",
+  year = "2023",
+  volume = "8",
+  number = "85",
+  pages = "5415",
+  doi = "10.21105/joss.05415"
+}
+
+@article{Page:2022,
+  url = {https://doi.org/10.1002/essoar.10510291.1},
+  title = "Access to Analysis and Climate Indices Tools for Climate Researchers and End Users",
+  author = "Page, Alessandro Spinuso, Lars Barring, Klaus Zimmermann, and Abel Aoun",
+  journal = "ESS Open Archive",
+  year = "2022",
+  doi = "10.1002/essoar.10510291.1",
+}
+
+@article{Brimicombe:2022,
+  url = {https://www.sciencedirect.com/science/article/pii/S2352711022000176},
+  title = "Thermofeel: A Python thermal comfort indices library",
+  author = "Brimicombe, C. and Di Napoli, C. and Quintino, T. and Pappenberger, F. and Cornforth, R. and Cloke, H. L.",
+  journal = "SoftwareX",
+  year = "2022",
+  volume = "18",
+  pages = "101005",
+  doi = "10.1016/j.softx.2022.101005"
 }
