+++ docs/source/contribution.rst
@@ -1,0 +1,69 @@
+.. raw:: html
+
+   <p>
+      <a href="https://github.com/DahyannAraya/copernicus-seasonal-forecast-tools">
+         <img src="https://img.shields.io/badge/github-repo-000.svg?logo=github&labelColor=gray&color=blue" alt="GitHub repo">
+      </a>
+      <a href="https://github.com/DahyannAraya/copernicus-seasonal-forecast-tools/blob/main/LICENSE">
+         <img src="https://img.shields.io/github/license/DahyannAraya/copernicus-seasonal-forecast-tools" alt="License">
+      </a>
+      <a href="https://badge.fury.io/py/copernicus-seasonal-forecast-tools">
+         <img src="https://badge.fury.io/py/copernicus-seasonal-forecast-tools.svg" alt="PyPI version">
+      </a>
+      <a href="https://pypi.org/project/copernicus-seasonal-forecast-tools/">
+         <img src="https://img.shields.io/badge/python-3.10–3.12-blue?logo=python&logoColor=white" alt="Supported Python versions">
+      </a>
+      <a href="https://copernicus-seasonal-forecast-tools.readthedocs.io/en/latest/?badge=latest">
+         <img src="https://readthedocs.org/projects/copernicus-seasonal-forecast-tools/badge/?version=latest" alt="Documentation Status">
+      </a>
+   </p>
+
+Contribution
+============
+
+This section summarizes how to contribute and where to find details. We adopt the CLIMADA contribution workflow and conventions.
+
+1) Contribute
+~~~~~~~~~~~~~
+
+We follow the CLIMADA workflow for branching, tests, and reviews. Please read the CLIMADA guide:
+
+- `CLIMADA development guide <https://climada-python.readthedocs.io/en/latest/development/index.html>`_
+
+Typical steps: fork -> branch -> tests -> pull request.
+
+For extended resources for CLIMADA development, visit:
+
+- `Guide to CLIMADA development <https://climada-python.readthedocs.io/en/latest/development/Guide_CLIMADA_Development.html>`_
+
+2) Report issues or problems
+----------------------------
+
+`Open an issue <https://github.com/DahyannAraya/copernicus-seasonal-forecast-tools/issues/new/choose>`_ and include:
+
+- Clear description and minimal example
+- Expected vs actual behavior
+- Version info (package, Python, OS)
+- Relevant logs or traceback
+
+3) Seek support
+---------------
+
+`Talk to the CLIMADA group <https://climada-python.readthedocs.io/en/latest/development/Guide_CLIMADA_Development.html#talk-to-the-group>`__
+
+- Before starting to code, coordinate with one of the repository admins.
+- Align on the big picture: intersections with other projects, potential conflicts, changes to CLIMADA core, and additional dependencies.
+- Also talk with members of the core development team (see the GitHub wiki).
+- Bring your idea to a developers meeting to get advice, find collaborators, and identify reviewers.
+- Keep communicating as your plans evolve.
+
+4) How to stay informed or participate
+--------------------------------------
+
+- `GitHub repository <https://github.com/CLIMADA-project/climada_python>`_
+- `Developer and user mailing list <https://sympa.ethz.ch/sympa/info/climada-users>`_
+- Monthly developers and users meetings (please register for the mailing list to receive invitations)
+- `Contribution guide <https://climada-python.readthedocs.io/en/latest/development/index.html>`_
+
+
+
