+++ paper.bib
@@ -169,33 +169,34 @@
 }
 
 @article{Bourgault:2023,
-  url = {https://joss.theoj.org/papers/10.21105/joss.05415},
-  title = "xclim: xarray-based climate data analytics",
-  author = "Bourgault, Pascal and Huard, David and Smith, Trevor James and Logan, Travis and others",
-  journal = "Journal of Open Source Software",
-  year = "2023",
-  volume = "8",
-  number = "85",
-  pages = "5415",
-  doi = "10.21105/joss.05415"
+  url = {https://doi.org/10.21105/joss.05415},
+  author = {Bourgault, Pascal and Huard, David and Smith, Trevor James and Logan, Travis and others},
+  title = {{xclim}: xarray-based climate data analytics},
+  journal = {Journal of Open Source Software},
+  year = {2023},
+  volume = {8},
+  number = {85},
+  pages = {5415},
+  doi = {10.21105/joss.05415}
 }
 
 @article{Page:2022,
   url = {https://doi.org/10.1002/essoar.10510291.1},
-  title = "Access to Analysis and Climate Indices Tools for Climate Researchers and End Users",
-  author = "Page, Alessandro Spinuso, Lars Barring, Klaus Zimmermann, and Abel Aoun",
-  journal = "ESS Open Archive",
-  year = "2022",
-  doi = "10.1002/essoar.10510291.1",
+  author = {Pagé, Christian and Spinuso, Alessandro and Bärring, Lars and Zimmermann, Klaus and Aoun, Abel},
+  title = {Access to Analysis and Climate Indices Tools for Climate Researchers and End Users},
+  journal = {ESS Open Archive},
+  year = {2022},
+  doi = {10.1002/essoar.10510291.1},
+  note = {Preprint}
 }
 
 @article{Brimicombe:2022,
-  url = {https://www.sciencedirect.com/science/article/pii/S2352711022000176},
-  title = "Thermofeel: A Python thermal comfort indices library",
-  author = "Brimicombe, C. and Di Napoli, C. and Quintino, T. and Pappenberger, F. and Cornforth, R. and Cloke, H. L.",
-  journal = "SoftwareX",
-  year = "2022",
-  volume = "18",
-  pages = "101005",
-  doi = "10.1016/j.softx.2022.101005"
+  url = {https://doi.org/10.1016/j.softx.2022.101005},
+  author = {Brimicombe, Chloe and Di Napoli, Claudia and Quintino, Tiago and Pappenberger, Florian and Cornforth, Rosalind and Cloke, Hannah L.},
+  title = {{Thermofeel}: A Python thermal comfort indices library},
+  journal = {SoftwareX},
+  year = {2022},
+  volume = {18},
+  pages = {101005},
+  doi = {10.1016/j.softx.2022.101005}
 }
