+++ paper.bib
@@ -153,9 +153,8 @@
 @misc{CDS,
   url = {https://cds.climate.copernicus.eu/},
   author = {{Copernicus Climate Change Service}},
-  title = {Description of the {C3S} seasonal multi-system},
+  title = {Copernicus Climate Data Store (CDS)},
   year = {2024},
-  howpublished = {\url{https://cds.climate.copernicus.eu/}},
   note = {Accessed: 2024-12-01}
 }
 
