+++ paper.md
@@ -54,7 +54,7 @@ Despite the growing interest in both seasonal prediction [@Osman:2023; @Ngoungue
 
 While existing libraries such as xclim [@Bourgault:2023], icclim [@Page:2022], and Thermofeel [@Brimicombe:2022] provide robust implementations of climate indices, they do not offer an operational pathway from seasonal forecasts to impact modeling frameworks. Our work complements these established tools by focusing specifically on the operational gap between seasonal climate predictions and risk assessment. We utilize Thermofeel as a reference for specific index calculations, ensuring our formulas align with published definitions to maintain comparability and facilitate reuse across the scientific community.
 
-The Copernicus Seasonal Forecast Tools Package addresses this gap by providing an automated, flexible, and modular solution. It connects seasonal forecast data from the Copernicus Climate Data Store (CDS) with the hazard and impact modeling capabilities of the CLIMADA platform [@AznarSiguan:2019], an open-source platform for climate risk analysis and the evaluation of adaptation benefits [@bresch_climada_2021]. This integration allows users to efficiently transform raw seasonal predictions into actionable climate risk insights. The package supports the definition of custom climate indices, the integration of data from multiple providers, and the adaptation to emerging hazard types and regional contexts. Thereby, it strengthens the operational implementation of impact forecasting and facilitates downstream components of the warning value chain or operational pipeline [@Golding:2019], bridging the gap between probabilistic seasonal forecasts and concrete impact assessments in an actionable timescale.
+The Copernicus Seasonal Forecast Tools Package addresses this gap by providing an automated, flexible, and modular solution. It connects seasonal forecast data from the Copernicus Climate Data Store [@CDS] with the hazard and impact modeling capabilities of the CLIMADA platform [@AznarSiguan:2019], an open-source platform for climate risk analysis and the evaluation of adaptation benefits [@bresch_climada_2021]. This integration allows users to efficiently transform raw seasonal predictions into actionable climate risk insights. The package supports the definition of custom climate indices, the integration of data from multiple providers, and the adaptation to emerging hazard types and regional contexts. Thereby, it strengthens the operational implementation of impact forecasting and facilitates downstream components of the warning value chain or operational pipeline [@Golding:2019], bridging the gap between probabilistic seasonal forecasts and concrete impact assessments in an actionable timescale.
 
 # Implementation and functionality
 
