### 2025-09-26 DahyannAraya/copernicus-seasonal-forecast-tools#11 @falconstryker
Improving usability
Overall, the functionality and performance of the software is great. It does what is described and I had no issues using it. Additionally, the parameters on [this page](https://copernicus-seasonal-forecast-tools.readthedocs.io/en/latest/climada_hazard_copernicus_forecast.html) are well-defined and items like `forecast.calculate_index(...)` and `forecast.download_data(...)` are also well-described. However, I suggest doing some cosmetic edits to improve readability and describing the results of these items more thoroughly. Specifically, things like:

> Example: index_metric = ClimateIndex.Tmax.name selects maximum temperature analysis.

would be improved as:

> Example: `index_metric = ClimateIndex.Tmax.name` selects maximum temperature analysis.

and this line appears as one long list when it might be better as a bulleted list:

> index_metric: Defines the type of index to be calculated. There are currently 12 predefined options available, including temperature-based indices (`Tmean` – Mean Temperature, `Tmin` – Minimum Temperature, `Tmax` – Maximum Temperature), heat stress indicators (`HIA` – Heat Index Adjusted, `HIS` – Heat Index Simplified, `HUM` – Humidex, `AT` – Apparent Temperature, `WBGT` – Wet Bulb Globe Temperature (Simple)), and extreme event indices (`HW` – Heat Wave, `TR` – Tropical Nights, `TX30` – Hot Days).

Also, some "customizing" instructions are vague, such as _"Users can define and integrate their own indices into the pipeline to extend the analysis according to their specific needs."_ This could use an example.

It would also be helpful to include a link pointing out what data specifically is downloaded by the `forecast. download_and_process_data` function.

The `forecast.calculate_index(...)` description under [Calculate Climate Indices](https://copernicus-seasonal-forecast-tools.readthedocs.io/en/latest/climada_hazard_copernicus_forecast.html#calculate-climate-indices) gives a good idea of what the process does, but not where to find the results. It took me awhile to discover that the data was stored in three files under:

`/Users/$USERNAME/climada/data/copernicus_data/seasonal_forecasts/dwd/sys21/2022/init11/valid12_02`

There is a lot of this throughout the walkthrough – somewhat incomplete instructions. I would love to see the Calculate Climate Indices section end with a summary of results and next steps, like:

"after running this command, you'll find the data under `/Users/blah/blah` in the folder `indices`. From here you can do X and Y with the data."

There should be code for reproducing the intensity plot as well. I cannot find the plot, if it's created at all.
