### 2025-09-26 #8827 @falconstryker
I finished my initial review. The functionality and performance of the software is great, however, the documentation needs improvement in order for the software to be useful. More on that [in this issue](https://github.com/DahyannAraya/copernicus-seasonal-forecast-tools/issues/11#issue-3459166278). I will update my checklist as the issues are addressed.

@elbeejay - the tests for this software are in a folder for running manually but not automated. I can see that that meets the "OK" standard for JOSS, and if that is sufficient for me to check it off of my list, please let me know.
