### 2025-09-26 #8827 @elbeejay
@falconstryker - we greatly appreciate the thorough review and constructive comments! To your question:

> @elbeejay - the tests for this software are in a folder for running manually but not automated. I can see that that meets the "OK" standard for JOSS, and if that is sufficient for me to check it off of my list, please let me know.

You are correct that it is "OK" per the JOSS-standard. You are welcome to make recommendations and suggestions for authors to make improvements. If testing exists that serves the purpose of software testing (i.e., verifies the functionality of the code), it is acceptable for JOSS even if it not automated. If testing appears to be incomplete or does not properly cover the functionality of the software presented, then it becomes an item that should be resolved prior to publication.
