### 2025-09-28 ad49187
tests: mocked integration, eliminate deep path dependency
