+++ test/test_integration_forecast_to_hazard.py
@@ -1,19 +1,28 @@
 """
-Integration tests for the SeasonalForecast pipeline in seasonal_forecast_tools.
+Integration tests for the SeasonalForecast pipeline using mocked/synthetic data.
 
 This test suite verifies that the full workflow executes correctly:
-from local GRIB input to NetCDF index computation and hazard generation (HDF5),
+from synthetic GRIB-like inputs to NetCDF index computation and hazard generation (HDF5),
 for all supported thermal index metrics.
 
-Each test checks file creation, dataset content, and hazard consistency.
+Uses mocked data to test integration without requiring deep repository structure.
+Completely restructured to eliminate dependency on physical files in repository.
 """
 
-from pathlib import Path
+import tempfile
 import unittest
-
+from itertools import cycle
+from pathlib import Path
+from unittest.mock import patch, MagicMock, mock_open, call
+import numpy as np
+import pandas as pd
 import xarray as xr
 
-from seasonal_forecast_tools.core.seasonal_forecast import SeasonalForecast, Hazard, CLIMADA_INSTALLED
+from seasonal_forecast_tools.core.seasonal_forecast import SeasonalForecast, CLIMADA_INSTALLED
+
+# If CLIMADA is available, import Hazard for testing
+if CLIMADA_INSTALLED:
+    from seasonal_forecast_tools.core.seasonal_forecast import Hazard  # noqa: F401
 
 INDEX_METRICS = [
     "Tmean",  # Mean Temperature
@@ -29,123 +38,513 @@ INDEX_METRICS = [
     "TX30"    # Hot Days
 ]
 
-class TestIntegrationWorkflow(unittest.TestCase):
+
+class TestIntegrationWorkflowMocked(unittest.TestCase):
     """
-    Integration test for the seasonal forecast pipeline:
-    From local GRIB data to index NetCDF to Hazard HDF5.
+    Integration test for the seasonal forecast pipeline using mocked data.
+    Tests the complete workflow without requiring repository file structure.
+    Completely restructured to eliminate physical file dependencies.
     """
 
-    @classmethod
-    def setUpClass(cls):
-        """
-        Prepare the test environment and run the full seasonal forecast pipeline.
+    def setUp(self):
+        """Set up temporary directory and mock data for each test."""
+        self.temp_dir = tempfile.TemporaryDirectory()
+        self.base_dir = Path(self.temp_dir.name)
 
-        This method initializes SeasonalForecast objects for each index metric using
-        predefined parameters, skips downloading (GRIB files exist locally),
-        processes the data, calculates the indices, saves hazards, and stores paths.
-        """
-        repo_root = Path(__file__).resolve().parents[1]
-        cls.base_dir = repo_root / "test" / "climada_data_test" / "copernicus_data" / "seasonal_forecasts"
+        # Test parameters matching original test but using mocks
+        self.originating_centre = "dwd"
+        self.system = "21"
+        self.year = 2022
+        self.init_month = 11
+        self.valid_period = ["December", "February"]
+        self.bounds = [-59, -35, -52, -29]  # W, S, E, N
 
-        cls.data_format = "grib"
-        cls.originating_centre = "dwd"
-        cls.system = "21"
-        cls.year = 2022
-        cls.init_month = 11
-        cls.valid_period = ["December", "February"]
-        cls.bounds = [-59, -35, -52, -29]  # W, S, E, N
+        # Store created forecasts for testing
+        self.forecasts = {}
+        self.mock_datasets = {}
 
-        cls.forecasts = {}
-        cls.hazard_paths = {}
-        cls.nc_paths = {}
+    def tearDown(self):
+        """Clean up temporary directory."""
+        self.temp_dir.cleanup()
+
[…]
