+++ test/test_path_construction.py
@@ -1,0 +1,25 @@
+"""
+Unit tests for path construction logic without requiring physical directories.
+Tests path building semantics using temporary directories and mocks.
+"""
+import unittest
+import tempfile
+from pathlib import Path
+from seasonal_forecast_tools.utils.path_utils import get_file_path
+
+class TestPathConstruction(unittest.TestCase):
+    def test_path_components_order(self):
+        """Test path components appear in correct order"""
+        with tempfile.TemporaryDirectory() as tmp_dir:
+            path = get_file_path(
+                tmp_dir, "dwd", 2023, "03", "06_08", 
+                "downloaded_data", "TR", "bounds", "21"
+            )
+            parts = Path(path).parts
+            self.assertIn("dwd", parts)
+            self.assertIn("sys21", parts)
+            self.assertIn("2023", parts)
+    
+    def test_filename_format(self):
+        """Test filename construction"""
+        # Test logic without deep structure

[…]
