### 2025-09-29 6b8adcf
add short example notebook
