### 2025-09-29 7d87471
adapt installation instructions
