### 2025-10-02 DahyannAraya/copernicus-seasonal-forecast-tools#7 @nishadhka
Hi @ValentinGebhart , well noted, this can be reflected in the part in the paper discussing the utils section.
