### 2025-10-04 #8827 @DahyannAraya
Hello @elbeejay,
Thank you for your message. We have addressed the majority of the open [issues](https://github.com/DahyannAraya/copernicus-seasonal-forecast-tools/issues) and are actively progressing on the remaining ones:

-	 [Issue #6 (Avoid sample inputs in deep nested folders)](https://github.com/DahyannAraya/copernicus-seasonal-forecast-tools/issues/6) – under co-author review:
We created a dedicated branch with an updated test set and simplified directory layout: [tests_updates branch](https://github.com/DahyannAraya/copernicus-seasonal-forecast-tools/tree/tests_updates). This is currently under review by the co-owners.
-	[Issue #7 (Using well known method in utils)](https://github.com/DahyannAraya/copernicus-seasonal-forecast-tools/issues/7) – responded:
The bounding_box_from_countries function is a thin wrapper around geopandas.GeoSeries.total_bounds, with pre/post-processing steps. We clarified this in the documentation.
-	 [Issue #8 (Two JOSS paper items)](https://github.com/DahyannAraya/copernicus-seasonal-forecast-tools/issues/8) – under co-author review:
I have already updated the paper.md on this regarding concern; currently, these changes are under review by the co-authors.
-	 [Issue #9 (readthedocs organization and clarity)](https://github.com/DahyannAraya/copernicus-seasonal-forecast-tools/issues/9) – responded:
We restructured the documentation according to the reviewer’s recommendations. The new structure separates installation, examples, walkthrough, and reference sections for improved clarity.
-	[Issue #10 (Broken links in readthedocs)](https://github.com/DahyannAraya/copernicus-seasonal-forecast-tools/issues/10) – responded:
We reviewed all the links and all work correctly across the package.
-	[Issue #11 (Improving usability)](https://github.com/DahyannAraya/copernicus-seasonal-forecast-tools/issues/11)  – under co-author review:
We reviewed the proposed changes and planned the implementations.

We are now waiting for feedback from the reviewers on some of the issues we have responded to, in order to reach agreement and close them. We plan to finalize the remaining items currently under internal review within the next few days. Once those updates are merged, the repository will reflect all requested changes. Most of the comments here are related to and expanded upon in the issues where we have attempted to respond in detail to the reviewer's concerns, comments, or requests. However, we will certainly double-check this week to see if there are any comments on this page that we have not addressed in the issues.

As always, we remain attentive to your comments.
