### 2025-10-06 5a9b432
improve description of tutorials
