### 2025-10-06 DahyannAraya/copernicus-seasonal-forecast-tools#11 @ValentinGebhart
Dear @falconstryker,

thanks a lot for all of these very helpful suggestions. We have added all of your suggestions to the new version of the [Full Walkthrough of the readthedocs](https://copernicus-seasonal-forecast-tools.readthedocs.io/en/latest/climada_hazard_copernicus_forecast.html).

Concerning your question about code to reproduce the intensity plot: The code should be visible in the [Full Walkthrough of the readthedocs](https://copernicus-seasonal-forecast-tools.readthedocs.io/en/latest/climada_hazard_copernicus_forecast.html) (section: Example for reading and plotting hazard). If you want to save the plot, you can use standard ways of saving a `GeoAxes` object.

Please let us know if the changes addressed your concerns. We definitely think that the changes improved the tutorial a lot!
