### 2025-10-06 6e404ff
added Thermofeel references
