+++ paper.md
@@ -62,7 +62,7 @@ The Copernicus Climate Data Store currently offers one of the most comprehensive
 
 To address this issue, the Copernicus Seasonal Forecast Package provides streamlined tools to access data via Climate Data Store API (Python client: cdsapi), and to filter and aggregate raw GRIB/netCDF files into daily netCDF outputs, facilitating their downstream use in climate index calculation and risk modeling. Once retrieved, it processes the raw files into gridded daily netCDF datasets, structured by forecast date, ensemble member, latitude, and longitude. In all steps of the process, the package checks if the corresponding output files already exist before proceeding. Each file includes multi-ensemble data for daily mean, maximum, and minimum values.
 
-Once daily data is calculated, users can select from twelve available heat indices, including temperature-based indices (Tmean, Tmin, Tmax), heat stress indicators (HIA—Heat Index Adjusted, HIS—Heat Index Simplified, HUM—Humidex, AT—Apparent Temperature, WBGT—Wet Bulb Globe Temperature), and extreme event indices (HW—Heat Wave, TR—Tropical Nights, TX30—Hot Days).
+Once daily data is calculated, users can select from twelve available heat indices (based on index definitions given in [@Brimicombe:2022]), including temperature-based indices (Tmean, Tmin, Tmax), heat stress indicators (HIA—Heat Index Adjusted, HIS—Heat Index Simplified, HUM—Humidex, AT—Apparent Temperature, WBGT—Wet Bulb Globe Temperature), and extreme event indices (HW—Heat Wave, TR—Tropical Nights, TX30—Hot Days).
 
 Calculated index data is saved in netCDF format and organized by index and time period, with file paths printed during processing. Statistical summaries of the index across ensemble members are computed and saved simultaneously, providing insight into forecast behavior. In addition to summary statistics, full access to all ensemble members is preserved to allow users to explore the full range of forecast uncertainty, rather than relying solely on aggregated values. This design choice supports transparent communication of uncertainty and more robust decision-making in impact-based applications. Users can then transform the index into a CLIMADA-compatible hazard object, enabling integration into impact-based forecasting workflows.
 
