### 2025-10-06 DahyannAraya/copernicus-seasonal-forecast-tools#9 @falconstryker
Thanks, the docs look great now. I think the restructuring works well and the docs are easier to follow.

I started to follow the instructions again and had issues with the install. First, in the install for PyPI, there's a step missing: either add `cd copernicus-seasonal-forecast-tools` step before `pip install` or revise `pip install` to be `pip install -r copernicus-seasonal-forecast-tools/docs/requirements.txt`

Second, I was unable to install climada. In both the **PyPI AND Install from GitHub** instructions, I had issues at the step to install with `requirements.txt`:

```
(climada) 12:25:$USER@$HOST:~>> pip install -r copernicus-seasonal-forecast-tools/docs/requirements.txt
Collecting git+https://github.com/CLIMADA-project/climada_python.git (from -r copernicus-seasonal-forecast-tools/docs/requirements.txt (line 1))
  Cloning https://github.com/CLIMADA-project/climada_python.git to /private/var/folders/fm/p9lm04096952ywr44vxs9p080000gq/T/pip-req-build-h5y5j67i
  Running command git clone --filter=blob:none --quiet https://github.com/CLIMADA-project/climada_python.git /private/var/folders/fm/p9lm04096952ywr44vxs9p080000gq/T/pip-req-build-h5y5j67i
  Resolved https://github.com/CLIMADA-project/climada_python.git to commit bb9595944fcf673152ea53e61efbafeb0b1e4406
  Installing build dependencies ... done
  Preparing metadata (pyproject.toml) ... done
Collecting fiona (from -r copernicus-seasonal-forecast-tools/docs/requirements.txt (line 2))
  Obtaining dependency information for fiona from https://files.pythonhosted.org/packages/65/0c/e8070b15c8303f60bd4444a120842597ccd6ed550548948e2e36cffbaa93/fiona-1.10.1-cp311-cp311-macosx_11_0_arm64.whl.metadata
  Using cached fiona-1.10.1-cp311-cp311-macosx_11_0_arm64.whl.metadata (56 kB)
Collecting pyarrow (from -r copernicus-seasonal-forecast-tools/docs/requirements.txt (line 3))
  Obtaining dependency information for pyarrow from https://files.pythonhosted.org/packages/94/dc/80564a3071a57c20b7c32575e4a0120e8a330ef487c319b122942d665960/pyarrow-21.0.0-cp311-cp311-macosx_12_0_arm64.whl.metadata
  Using cached pyarrow-21.0.0-cp311-cp311-macosx_12_0_arm64.whl.metadata (3.3 kB)
Collecting bayesian-optimization (from climada==6.1.0->-r copernicus-seasonal-forecast-tools/docs/requirements.txt (line 1))
  Obtaining dependency information for bayesian-optimization from https://files.pythonhosted.org/packages/2f/7b/894dc80becb0cde4e099d603a6064d0e1acd26f81a933fdc443e1b664968/bayesian_optimization-3.1.0-py3-none-any.whl.metadata
  Using cached bayesian_optimization-3.1.0-py3-none-any.whl.metadata (11 kB)
Collecting bottleneck (from climada==6.1.0->-r copernicus-seasonal-forecast-tools/docs/requirements.txt (line 1))
  Obtaining dependency information for bottleneck from https://files.pythonhosted.org/packages/83/96/9d51012d729f97de1e75aad986f3ba50956742a40fc99cbab4c2aa896c1c/bottleneck-1.6.0-cp311-cp311-macosx_11_0_arm64.whl.metadata
  Using cached bottleneck-1.6.0-cp311-cp311-macosx_11_0_arm64.whl.metadata (8.2 kB)
Requirement already satisfied: cartopy in ./climada/lib/python3.11/site-packages (from climada==6.1.0->-r copernicus-seasonal-forecast-tools/docs/requirements.txt (line 1)) (0.25.0)
Requirement already satisfied: cfgrib in ./climada/lib/python3.11/site-packages (from climada==6.1.0->-r copernicus-seasonal-forecast-tools/docs/requirements.txt (line 1)) (0.9.15.1)
Collecting contextily (from climada==6.1.0->-r copernicus-seasonal-forecast-tools/docs/requirements.txt (line 1))
  Obtaining dependency information for contextily from https://files.pythonhosted.org/packages/fb/46/07a029b73f9a5c7bbf9b538e6441c42014a448f335a1cc780616f2594bad/contextily-1.6.2-py3-none-any.whl.metadata
  Using cached contextily-1.6.2-py3-none-any.whl.metadata (2.9 kB)
Collecting dask (from climada==6.1.0->-r copernicus-seasonal-forecast-tools/docs/requirements.txt (line 1))
  Obtaining dependency information for dask from https://files.pythonhosted.org/packages/25/60/3fcd548bed6d25016933e4b2984c9b82e4c1e760380e03d4100b1b4726e0/dask-2025.9.1-py3-none-any.whl.metadata
  Using cached dask-2025.9.1-py3-none-any.whl.metadata (3.8 kB)
Collecting deprecation (from climada==6.1.0->-r copernicus-seasonal-forecast-tools/docs/requirements.txt (line 1))
  Obtaining dependency information for deprecation from https://files.pythonhosted.org/packages/02/c3/253a89ee03fc9b9682f1541728eb66db7db22148cd94f89ab22528cd1e1b/deprecation-2.1.0-py2.py3-none-any.whl.metadata
  Using cached deprecation-2.1.0-py2.py3-none-any.whl.metadata (4.6 kB)
Collecting gdal (from climada==6.1.0->-r copernicus-seasonal-forecast-tools/docs/requirements.txt (line 1))
  Using cached gdal-3.11.4.tar.gz (878 kB)
  Installing build dependencies ... done
  Getting requirements to build wheel ... error
  error: subprocess-exited-with-error

  × Getting requirements to build wheel did not run successfully.
  │ exit code: 1
  ╰─> [96 lines of output]
      /private/var/folders/fm/p9lm04096952ywr44vxs9p080000gq/T/pip-build-env-d3q8qbr1/overlay/lib/python3.11/site-packages/setuptools/config/_apply_pyprojecttoml.py:82: SetuptoolsWarning: `license` overwritten by `pyproject.toml`
        corresp(dist, value, root_dir)
      /private/var/folders/fm/p9lm04096952ywr44vxs9p080000gq/T/pip-build-env-d3q8qbr1/overlay/lib/python3.11/site-packages/setuptools/config/_apply_pyprojecttoml.py:82: SetuptoolsWarning: `extras_require` overwritten in `pyproject.toml` (optional-dependencies)
        corresp(dist, value, root_dir)
      Using numpy 2.3.3
      running egg_info
      writing gdal-utils/GDAL.egg-info/PKG-INFO
      writing dependency_links to gdal-utils/GDAL.egg-info/dependency_links.txt
      writing entry points to gdal-utils/GDAL.egg-info/entry_points.txt
      writing requirements to gdal-utils/GDAL.egg-info/requires.txt
      writing top-level names to gdal-utils/GDAL.egg-info/top_level.txt
      Traceback (most recent call last):
        File "<string>", line 92, in fetch_config
        File "/Users/$USER/opt/anaconda3/lib/python3.11/subprocess.py", line 1026, in __init__
          self._execute_child(args, executable, preexec_fn, close_fds,
        File "/Users/$USER/opt/anaconda3/lib/python3.11/subprocess.py", line 1950, in _execute_child
          raise child_exception_type(errno_num, err_msg, err_filename)
      FileNotFoundError: [Errno 2] No such file or directory: 'gdal-config'

      Traceback (most recent call last):
        File "<string>", line 244, in get_gdal_config
        File "<string>", line 95, in fetch_config
      gdal_config_error: [Errno 2] No such file or directory: 'gdal-config'

      Traceback (most recent call last):
        File "/Users/$USER/climada/lib/python3.11/site-packages/pip/_vendor/pyproject_hooks/_in_process/_in_process.py", line 353, in <module>
          main()
        File "/Users/$USER/climada/lib/python3.11/site-packages/pip/_vendor/pyproject_hooks/_in_process/_in_process.py", line 335, in main
          json_out['return_val'] = hook(**hook_input['kwargs'])
                                   ^^^^^^^^^^^^^^^^^^^^^^^^^^^^
        File "/Users/$USER/climada/lib/python3.11/site-packages/pip/_vendor/pyproject_hooks/_in_process/_in_process.py", line 118, in get_requires_for_build_wheel
          return hook(config_settings)
                 ^^^^^^^^^^^^^^^^^^^^^
        File "/private/var/folders/fm/p9lm04096952ywr44vxs9p080000gq/T/pip-build-env-d3q8qbr1/overlay/lib/python3.11/site-packages/setuptools/build_meta.py", line 331, in get_requires_for_build_wheel
          return self._get_build_requires(config_settings, requirements=[])
                 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
        File "/private/var/folders/fm/p9lm04096952ywr44vxs9p080000gq/T/pip-build-env-d3q8qbr1/overlay/lib/python3.11/site-packages/setuptools/build_meta.py", line 301, in _get_build_requires
          self.run_setup()
        File "/private/var/folders/fm/p9lm04096952ywr44vxs9p080000gq/T/pip-build-env-d3q8qbr1/overlay/lib/python3.11/site-packages/setuptools/build_meta.py", line 317, in run_setup
          exec(code, locals())
        File "<string>", line 456, in <module>
        File "/private/var/folders/fm/p9lm04096952ywr44vxs9p080000gq/T/pip-build-env-d3q8qbr1/overlay/lib/python3.11/site-packages/setuptools/__init__.py", line 115, in setup
          return distutils.core.setup(**attrs)
                 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
        File "/private/var/folders/fm/p9lm04096952ywr44vxs9p080000gq/T/pip-build-env-d3q8qbr1/overlay/lib/python3.11/site-packages/setuptools/_distutils/core.py", line 186, in setup
          return run_commands(dist)
                 ^^^^^^^^^^^^^^^^^^
        File "/private/var/folders/fm/p9lm04096952ywr44vxs9p080000gq/T/pip-build-env-d3q8qbr1/overlay/lib/python3.11/site-packages/setuptools/_distutils/core.py", line 202, in run_commands
          dist.run_commands()
        File "/private/var/folders/fm/p9lm04096952ywr44vxs9p080000gq/T/pip-build-env-d3q8qbr1/overlay/lib/python3.11/site-packages/setuptools/_distutils/dist.py", line 1002, in run_commands
          self.run_command(cmd)
        File "/private/var/folders/fm/p9lm04096952ywr44vxs9p080000gq/T/pip-build-env-d3q8qbr1/overlay/lib/python3.11/site-packages/setuptools/dist.py", line 1102, in run_command
          super().run_command(command)
        File "/private/var/folders/fm/p9lm04096952ywr44vxs9p080000gq/T/pip-build-env-d3q8qbr1/overlay/lib/python3.11/site-packages/setuptools/_distutils/dist.py", line 1021, in run_command
          cmd_obj.run()
        File "/private/var/folders/fm/p9lm04096952ywr44vxs9p080000gq/T/pip-build-env-d3q8qbr1/overlay/lib/python3.11/site-packages/setuptools/command/egg_info.py", line 312, in run
          self.find_sources()
        File "/private/var/folders/fm/p9lm04096952ywr44vxs9p080000gq/T/pip-build-env-d3q8qbr1/overlay/lib/python3.11/site-packages/setuptools/command/egg_info.py", line 320, in find_sources
          mm.run()
        File "/private/var/folders/fm/p9lm04096952ywr44vxs9p080000gq/T/pip-build-env-d3q8qbr1/overlay/lib/python3.11/site-packages/setuptools/command/egg_info.py", line 543, in run
          self.add_defaults()
        File "/private/var/folders/fm/p9lm04096952ywr44vxs9p080000gq/T/pip-build-env-d3q8qbr1/overlay/lib/python3.11/site-packages/setuptools/command/egg_info.py", line 581, in add_defaults
          sdist.add_defaults(self)
        File "/private/var/folders/fm/p9lm04096952ywr44vxs9p080000gq/T/pip-build-env-d3q8qbr1/overlay/lib/python3.11/site-packages/setuptools/command/sdist.py", line 109, in add_defaults
          super().add_defaults()
        File "/private/var/folders/fm/p9lm04096952ywr44vxs9p080000gq/T/pip-build-env-d3q8qbr1/overlay/lib/python3.11/site-packages/setuptools/_distutils/command/sdist.py", line 245, in add_defaults
          self._add_defaults_ext()
        File "/private/var/folders/fm/p9lm04096952ywr44vxs9p080000gq/T/pip-build-env-d3q8qbr1/overlay/lib/python3.11/site-packages/setuptools/_distutils/command/sdist.py", line 329, in _add_defaults_ext
          build_ext = self.get_finalized_command('build_ext')
                      ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
        File "/private/var/folders/fm/p9lm04096952ywr44vxs9p080000gq/T/pip-build-env-d3q8qbr1/overlay/lib/python3.11/site-packages/setuptools/_distutils/cmd.py", line 334, in get_finalized_command
          cmd_obj.ensure_finalized()
        File "/private/var/folders/fm/p9lm04096952ywr44vxs9p080000gq/T/pip-build-env-d3q8qbr1/overlay/lib/python3.11/site-packages/setuptools/_distutils/cmd.py", line 135, in ensure_finalized
          self.finalize_options()
        File "<string>", line 328, in finalize_options
        File "<string>", line 250, in get_gdal_config
      gdal_config_error: Traceback (most recent call last):
        File "<string>", line 92, in fetch_config
        File "/Users/$USER/opt/anaconda3/lib/python3.11/subprocess.py", line 1026, in __init__
          self._execute_child(args, executable, preexec_fn, close_fds,
        File "/Users/$USER/opt/anaconda3/lib/python3.11/subprocess.py", line 1950, in _execute_child
          raise child_exception_type(errno_num, err_msg, err_filename)
      FileNotFoundError: [Errno 2] No such file or directory: 'gdal-config'

      Traceback (most recent call last):
        File "<string>", line 244, in get_gdal_config
        File "<string>", line 95, in fetch_config
      gdal_config_error: [Errno 2] No such file or directory: 'gdal-config'

      Could not find gdal-config. Make sure you have installed the GDAL native library and development headers.
      [end of output]

error: subprocess-exited-with-error

× Getting requirements to build wheel did not run successfully.
│ exit code: 1
╰─> See above for output.

```

I also couldn't install climada alone (the step with `git clone https://github.com/CLIMADA-project/climada_python.git`).

If you could advise/fix the install issue, I'll continue!
