### 2025-10-06 #8827 @falconstryker
> > [@falconstryker](https://github.com/falconstryker) - both are great questions. I agree with your concern that it is important we clarify which repository is the primary one for the purposes of reviewing, citing, and associating with the JOSS publication. I'd also be curious to learn why it is a "fork", but also not a true GitHub fork.
> > We do permit authorship without requiring contributions to the code (per [JOSS policy](https://joss.readthedocs.io/en/latest/submitting.html#authorship)), however it is entirely valid to ask about co-author contributions; this kind of question does come up from time to time.
> > I think for both questions, [@DahyannAraya](https://github.com/DahyannAraya) will need to regply.
>
> Dear [@falconstryker](https://github.com/falconstryker) and [@elbeejay](https://github.com/elbeejay),
>
> thank you for these questions, and sorry for this late reply. Here some info to clarify the authors and the repository structure.
>
> * This package, code, and functionalities were conceptualized as part of the [U-CLIMADAPT](https://www.dwd.de/EN/research/projects/socioeconomics/fpcup_u_climadapt/fpcup_u_climadapt.html) project by David Bresch (ETH Zurich), Tobias Geiger (DWD, German weather service) and Dahyann Araya (ETH Zurich).
> * We first started to implement this package in [CLIMADA petals](https://github.com/CLIMADA-project/climada_petals) -- see this [github history](https://github.com/CLIMADA-project/climada_petals/compare/main...feature/copernicus_forecast) for initial commits -- which is part of the CLIMADA code base but not part of CLIMADA core as it is more about data generation.
> * As soon as we realized that this package has value beyond CLIMADA and that we did not want it to be too dependent on CLIMADA, we decided to create a separate package (that is still easily coupled to CLIMADA if wanted).
> * We anyway decided to have a [clone](https://github.com/CLIMADA-project/copernicus-seasonal-forecast-tools) in the CLIMADA-project repository, as the package was first developed as a submodule of CLIMADA and should be clearly visible there. We also thought about having the CLIMADA-project clone a true GitHub fork and might implement that in the future.
>
> So, to concretely respond to your questions:
>
> 1. Yes, as was indeed done by the reviewers, we would like to have the official review (issues etc) on DahyannAeaya's [github repo](https://github.com/DahyannAraya/copernicus-seasonal-forecast-tools).
> 2. Tobias Geiger and David Bresch worked on this package mainly in the conceptualization phase and in several feedback iterations (also when writing the paper). Thus they should be coauthors in the JOSS paper.
>
> I hope this clarified your questions, and sorry for the confusion!

So is the repo under @DahyannAraya [here](https://github.com/DahyannAraya/copernicus-seasonal-forecast-tools) the same code as the repo under Climada-project [here](https://github.com/CLIMADA-project/copernicus-seasonal-forecast-tools)? If not, should they be the same? Should the Climada-project one be deleted? This is something that @elbeejay may have to weigh in on since it sounds like a copyright/legal issue to have 1. two repos of the same name that are not true forks or 2. two different repos with the same name.

In either case, it is confusing. As a user, I would think to use the Climada-project version of copernicus-seasonal-forecast-tools because it sounds like it would be the "official" one to use, rather than the one hosted on someone's personal repository. To address that, I think you'd need to remove it from Climada-project or somehow have Climada-project points to @DahyannAraya's repo.
