### 2025-10-07 2c2a675
include cd to installation instructions
