### 2025-10-07 DahyannAraya/copernicus-seasonal-forecast-tools#9 @ValentinGebhart
Dear @falconstryker,

thanks a lot for making us aware of this issue! Indeed, also for us the current installation instructions do not work anymore. We will try to fix this ASAP, but it might take a bit of time as the colleague who made the instruciton files is on vacation.

We will let you know as soon as we fixed the problem!
