### 2025-10-07 #8827 @ValentinGebhart
> So is the repo under [@DahyannAraya](https://github.com/DahyannAraya) [here](https://github.com/DahyannAraya/copernicus-seasonal-forecast-tools) the same code as the repo under Climada-project [here](https://github.com/CLIMADA-project/copernicus-seasonal-forecast-tools)? If not, should they be the same? Should the Climada-project one be deleted? This is something that [@elbeejay](https://github.com/elbeejay) may have to weigh in on since it sounds like a copyright/legal issue to have 1. two repos of the same name that are not true forks or 2. two different repos with the same name.
>
> In either case, it is confusing. As a user, I would think to use the Climada-project version of copernicus-seasonal-forecast-tools because it sounds like it would be the "official" one to use, rather than the one hosted on someone's personal repository. To address that, I think you'd need to remove it from Climada-project or somehow have Climada-project points to [@DahyannAraya](https://github.com/DahyannAraya)'s repo.

Dear @falconstryker,

after internal discussion, we agree that the current structure is not optimal. We will soon replace the code copy in CLIMADA-project with a true GitHub fork pointing to the principal code base at [@DahyannAraya](https://github.com/DahyannAraya)'s repo. This might take a bit as our colleague who has the permissions to change CLIMADA-project is currently on vacation.

Hope that this solution makes sense, and thanks again for bringing it up!
