### 2025-11-06 f6fe75d
updated installation instruction in readthedocs
