### 2025-11-06 f55d7da
update installation readthedocs
