### 2025-11-06 8ac238f
update readthedocs installation
