### 2025-11-07 1306ba7
finalize installation instruction readthedocs
