+++ docs/source/installation.rst
@@ -29,7 +29,7 @@ Below we describe the installation using the package manager and environment man
 `Conda <https://docs.conda.io/projects/conda/en/stable/user-guide/getting-started.html>`_
 (similar installations are possible with the package managers `Mamba <https://mamba.readthedocs.io/en/latest/>`_ or `Pixi <https://pixi.sh/latest/>`_).
 
-To **install the package with the climate-risk assessment package CLIMADA**:
+To install the package **with** the climate-risk assessment package CLIMADA:
 
 .. code-block:: bash
 
@@ -37,7 +37,7 @@ To **install the package with the climate-risk assessment package CLIMADA**:
    conda activate conernicus_with
    pip install copernicus-seasonal-forecast-tools
 
-To **install the package without the climate-risk assessment package CLIMADA**:
+To install the package **without** the climate-risk assessment package CLIMADA:
 
 .. code-block:: bash
 
@@ -45,81 +45,28 @@ To **install the package without the climate-risk assessment package CLIMADA**:
    conda activate conernicus_without
    pip install copernicus-seasonal-forecast-tools
 
-To install the package **in developer (editable) mode on macOS/Linux**, and run the documentation and tests:
+To install the package **in developer (editable) mode**, and run the documentation and tests:
 
 .. code-block:: bash
 
+   conda create -n copernicus-dev-mode python=3.11 pip geopandas climada
+   conda activate copernicus-dev-mode
    git clone https://github.com/DahyannAraya/copernicus-seasonal-forecast-tools.git
    cd copernicus-seasonal-forecast-tools
-   python3.10 -m venv .venv_forecast
-   source .venv_forecast/bin/activate
    pip install -e .
-   pip install -r docs/requirements.txt # optional (including CLIMADA installation)
-
-To install the package **in developer (editable) mode on Windows (PowerShell)**, and run the documentation and tests:
-
-.. code-block:: powershell
-
-   git clone https://github.com/DahyannAraya/copernicus-seasonal-forecast-tools.git
-   cd copernicus-seasonal-forecast-tools
-   py -3.10 -m venv .venv_forecast
-   .venv_forecast\Scripts\Activate.ps1
-   pip install -e .
-   pip install -r docs/requirements.txt # optional (including CLIMADA installation)
-
-INSTALLATIONS TO BE REMOVED
-===========================
-
-Install from PyPI:
-
-.. code-block:: bash
-
-   pip install copernicus-seasonal-forecast-tools
-   # optional (including CLIMADA installation):
-   git clone https://github.com/DahyannAraya/copernicus-seasonal-forecast-tools.git
-   cd copernicus-seasonal-forecast-tools
-   pip install -r docs/requirements.txt
-
-INSTALLATION TO BE REMOVED
-.. OLD INSTALLATION INSTRUCTIONS; NOT FUNCTIONAL AT THE MOMENT
-Install via environment.yml (Conda or Mamba):
-
-.. code-block:: bash
-
-   git clone https://github.com/DahyannAraya/copernicus-seasonal-forecast-tools.git
-   cd copernicus-seasonal-forecast-tools
-   conda env create -f environment.yml
-   conda activate venv_forecast
-
-INSTALLATION TO BE REMOVED
-Install from GitHub:
-
-.. code-block:: bash
-
-   git clone https://github.com/DahyannAraya/copernicus-seasonal-forecast-tools.git
-   cd copernicus-seasonal-forecast-tools
-   pip install .
-   pip install -r docs/requirements.txt  # optional (including CLIMADA installation)
-
 
 CLIMADA Installation
 --------------------
 
 CLIMADA is required to generate hazard layers.
 
-- If you installed via `environment.yml`, CLIMADA is already included.
[…]
