### 2025-11-26 b4c4a71
adapt installation instructions in readme
