### 2025-11-26 b31af14
adapt reamde
