### 2025-11-26 a613363
adapt index.rst
