### 2025-12-02 46b6009
fix failing tests
