### 2025-12-02 DahyannAraya/copernicus-seasonal-forecast-tools#9 @ValentinGebhart
Dear @falconstryker, apologies for the late response!

There were indeed some installation issues with gdal that we could not solve without using conda/mamba. That is why we now have adapted the installation instructions. The new installation options now all work (at least on our devices). Thanks for the patience.
