### 2025-12-02 #8827 @ValentinGebhart
Dear @elbeejay, Dear @falconstryker, Dear @nishadhka,

apologies for the delayed response, unfortunately we had a busy couple of months and some of the noted issues turned out to have taken longer to fix than expected. Thank you a lot for your patience.

We have now responded to all comments in this thread and the [issues](https://github.com/DahyannAraya/copernicus-seasonal-forecast-tools/issues) in the repository. Please let us know if there are any remaining points or issues left to clarify.

We are looking forward to your further review and very helpful comments!
