### 2026-01-08 #8827 @elbeejay
> I'm satisfied with the changes and recommend publication.

Thank you @DamienIrving - do you mind ticking off the last checkbox in your [reviewer checklist](https://github.com/openjournals/joss-reviews/issues/8827#issuecomment-3218798277) when you get a chance? We ask that those all be checked off for accountability if we were to ever look back and revisit the review issue.
