+++ pyproject.toml
@@ -7,7 +7,7 @@ name = "copernicus-seasonal-forecast-tools"
 version = "0.1.2"
 description = "CLIMADA-compatible module for generating and analyzing seasonal forecast hazards from Copernicus data"
 license = "GPL-3.0-or-later"
-requires-python = ">=3.10,<3.12"
+requires-python = ">=3.10,<3.13"
 dependencies = [
     "xarray",
     "cfgrib",
