### 2026-01-22 DahyannAraya/copernicus-seasonal-forecast-tools#8 @falconstryker
The changes are sufficient and this issue can be closed
