### 2026-01-22 DahyannAraya/copernicus-seasonal-forecast-tools#11 @falconstryker
This issue is addressed and can be closed
