### 2026-01-22 #8827 @falconstryker
@elbeejay I have finished my review. The documentation is MUCH improved and the paper meets the standards. I recommended including installation instructions for pip (non-conda) since conda is limiting its user rate limit. Once that is resolved, I recommend for publication.
