### 2026-01-23 #8827 @elbeejay
Fantastic, thanks @falconstryker - from what I can tell, the package is hosted on PyPI and pip-installation is supported. Please confirm @ValentinGebhart
