### 2026-01-27 069184f
add conda-forge in installation instructions
