+++ docs/source/installation.rst
@@ -33,7 +33,7 @@ To install the package **with** the climate-risk assessment package CLIMADA:
 
 .. code-block:: bash
 
-   conda create -n copernicus_with python=3.11 pip climada
+   conda create -c conda-forge -n copernicus_with python=3.11 pip climada
    conda activate copernicus_with
    pip install copernicus-seasonal-forecast-tools
 
@@ -41,7 +41,7 @@ To install the package **without** the climate-risk assessment package CLIMADA:
 
 .. code-block:: bash
 
-   conda create -n copernicus_without python=3.11 pip geopandas
+   conda create -c conda-forge -n copernicus_without python=3.11 pip geopandas
    conda activate copernicus_without
    pip install copernicus-seasonal-forecast-tools
 
@@ -49,7 +49,7 @@ To install the package **in developer (editable) mode**, and run the documentati
 
 .. code-block:: bash
 
-   conda create -n copernicus-dev-mode python=3.11 pip geopandas climada
+   conda create -c conda-forge -n copernicus-dev-mode python=3.11 pip geopandas climada
    conda activate copernicus-dev-mode
    git clone https://github.com/DahyannAraya/copernicus-seasonal-forecast-tools.git
    cd copernicus-seasonal-forecast-tools
