+++ docs/source/index.rst
@@ -62,7 +62,7 @@ For a quick start, install the package and its requirements
 
 .. code-block:: bash
 
-   conda create -n copernicus_with python=3.11 pip climada
+   conda create -c conda-forge -n copernicus_with python=3.11 pip climada
    conda activate copernicus_with
    pip install copernicus-seasonal-forecast-tools
 
