### 2026-01-27 #8827 @ValentinGebhart
Dear @elbeejay and @falconstryker
Thanks a lot for your review. We responded to the installation comment in the [issue](https://github.com/DahyannAraya/copernicus-seasonal-forecast-tools/issues/9). In short: yes, the package is hosted on PyPI. Unfortunately, the package climada is not easily pip installable (due to its dependency gdal), which is why we use conda for the installation instructions.
