### 2026-01-31 #8827 @elbeejay
Great, it looks like all outstanding issues and comments have been resolved. Many, many thanks to our dear peer-reviewers @falconstryker, @DamienIrving, @nishadhka!

@ValentinGebhart, we have a few more editorial steps to go through to make sure your submission is ready for publication. I will be generating a small checklist with those items. Please just reply with a comment supplying me with some of the information I'll need (confirmation of latest version number, link to code archive, etc.)
