### 2026-02-12 #8827 @elbeejay
Hi @ValentinGebhart - please get back to me with the latest version number you'd like to associate with the publication as well as the DOI of the code archive (e.g., Zenodo) you set up. Thanks!
