### 2026-02-12 #8827 @ValentinGebhart
Dear @elbeejay, sorry for our delayed responses. We should be all set now:

- Authors, affiliations, and orchidIDs of the article are correct
- We made a github release with the latest changes. The version number (tag) is v0.1.1
- We made a Zenodo release with the correct author list and license, with this link: https://doi.org/10.5281/zenodo.18605471

Please let me know if further action from our side is needed. Thanks a lot!
