### 2026-02-15 #8827 @elbeejay
@ValentinGebhart are you able to edit the title of the Zenodo archive so it matches the JOSS paper title exactly? Right now it is "DahyannAraya/copernicus-seasonal-forecast-tools: copernicus-seasonal-forecast-tools v0.1.1 — JOSS submission" we would want it to be "Copernicus Seasonal Forecast Tools Package: Bridging Seasonal Climate Predictions and Impact Models for Operational Risk Assessment" if possible.
