### 2026-02-16 #8827 @ValentinGebhart
@elbeejay I adapted the title of the zenodo archive to match the paper title. Please let me know if there are more changes necessary.
