### 2026-02-16 #8827 @elbeejay
Thank you @ValentinGebhart, everything looks to be in order, one of the JOSS EiCs will do a fresh look over the materials and is responsible for the official publication.

Many many thanks again to our reviewers: @falconstryker, @DamienIrving, and @nishadhka - JOSS is not possible without folks like yourselves.
