### 2026-02-16 #8827 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
