### 2026-02-20 #8827 @kthyng
@DahyannAraya Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file. For example "copernicus" and maybe "python" but please check for others.
