### 2026-02-24 #8827 @ValentinGebhart
Dear @kthyng,
thanks for noting the wrong capitalization in the references. We have updated the bib file, the titles should now be displayed correctly.
