### 2026-03-06 #8827 @kthyng
Congratulations on your new publication @DahyannAraya! Many thanks to editor @elbeejay and to reviewers @falconstryker, @DamienIrving, @nishadhka for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
